main/requirements.txt
flask
pyyaml
flask-wtf
slacker
redis
